Print out time until next drink in a human friendly form.
Previously the time until the next drink was always printed in seconds.
E.g. 1s, 60s, 120s, 154564654s,

Now it is printed to the nearest second, minute, hour, or day using the
floor method.  Thus:
 1s -> 1s
 60s -> 1m
 120s -> 2m
 179s -> 2m
 3600s -> 1h

Admittedly, this looks a bit funny for values like 179.  But the idea is
to give a rough sense.  Could consider swapping floor semantics with
rounding, although it still has the same fundamental problem.